Add files via upload
the basics of HTML, CSS and JavaScript and fiddled around with some of the notable features of each of these languages. In the upcoming project, you will be creating the front-end of a blogging website using HTML, CSS and JavaScript. This blogging website enables one to register a new user, create his/her posts, edit an already existing post, add comments on a post, like a post, etc.